code/src/requirements.txt
streamlit
langchain
langchain-groq
langchain-openai
langchain-community
pysnc
requests
python-dotenv
langgraph
requests-oauthlib